move src back to root directory
