+++ src/util/stripHtmlTags.js
