+++ src/util/transform-token.js
