+++ webpack.prod.js
@@ -1,0 +1,7 @@
+const merge = require('webpack-merge');
+const baseConfig = require('./webpack.common');
+
+module.exports = merge(baseConfig, {
+  mode: 'production',
+  devtool: 'source-map'
+});
